Update password and avatar
