+++ actions/user.ts
@@ -3,8 +3,10 @@
 import { auth } from '@/lib/auth'
 import prisma from '@/lib/db'
 import { updateAvatarSchema, updateNameSchema, updatePasswordSchema } from '@/validators/user'
-import { getErrorRedirect, getSuccessRedirect, parseFormData } from '@cgambrell/utils'
+import { getSuccessRedirect, parseFormData } from '@cgambrell/utils'
+import bcrypt from 'bcryptjs'
 import { redirect } from 'next/navigation'
+import { uploadFile } from './s3'
 
 export async function updateName(formData: FormData) {
 	const { data, errors } = parseFormData(formData, updateNameSchema)
@@ -19,37 +21,22 @@ export async function updateName(formData: FormData) {
 export async function updateAvatar(formData: FormData) {
 	const { data, errors } = parseFormData(formData, updateAvatarSchema)
 	if (errors) return { errors }
-	// TODO: Can be in validator
-	else if (data.avatar.size === 0) redirect(getErrorRedirect('/settings', 'File must not be empty'))
 
 	const user = await auth()
 
-	const file = data.avatar
-	const fileExt = file.name.split('.').pop()
-	const filePath = `${user.id}/${new Date().getTime()}-${Math.random()}.${fileExt}`
-
-	// BUG: Not uploading to S3
-	// const { error: uploadError } = await supabase.storage.from('avatars').upload(filePath, file)
-	// if (uploadError) redirect(getErrorRedirect('/settings', uploadError.message))
-
-	// const { data: uploadedAvatar } = await supabase.storage.from('avatars').getPublicUrl(filePath)
-
-	// await prisma.user.update({where: {id: user.id}, data: {image: uploadedAvatar.publicUrl}})
+	const url = await uploadFile(data.avatar, `avatars/${user.id}`)
+	await prisma.user.update({ where: { id: user.id }, data: { image: url } })
 
 	redirect(getSuccessRedirect('/settings', 'Avatar updated'))
 }
 
-// BUG: Need to do this with prisma
 export async function updatePassword(_prevState: any, formData: FormData) {
 	const { data, errors } = parseFormData(formData, updatePasswordSchema)
 	if (errors) return { errors }
 
-	// 	if (!!data.password && data.password !== data.confirm_password) return { errors: { confirm_password: ['Passwords do not match'] } }
+	const user = await auth()
+	const passwordHash = await bcrypt.hash(data.password, 10)
+	await prisma.user.update({ where: { id: user.id }, data: { passwordHash } })
 
-	// 	const supabase = createClient()
-
-	// 	const { error } = await supabase.auth.updateUser({ password: data.password })
-	// 	if (error) redirect(getErrorRedirect('/settings/password', error.message))
-
-	redirect(getSuccessRedirect('/settings/password', 'Password updated'))
+	redirect(getSuccessRedirect('/settings/security', 'Password updated'))
 }
